MyGameDiary/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
Django==4.2
django-bootstrap-v5==1.0.11
django-extensions==3.2.3
exceptiongroup==1.2.1
executing==2.0.1
idna==3.7
ipython==8.24.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.0
urllib3==2.2.1
wcwidth==0.2.13
